refactor(MemberCard): 重构会员卡bean包下的包名
