+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/api/WxMpMemberCardService.java
@@ -1,13 +1,13 @@
 package me.chanjar.weixin.mp.api;
 
 import me.chanjar.weixin.common.exception.WxErrorException;
-import me.chanjar.weixin.mp.bean.mambercard.WxMpMemberCardActivatedMessage;
-import me.chanjar.weixin.mp.bean.mambercard.WxMpMemberCardUserInfoResult;
+import me.chanjar.weixin.mp.bean.membercard.WxMpMemberCardActivatedMessage;
+import me.chanjar.weixin.mp.bean.membercard.WxMpMemberCardUserInfoResult;
 
 /**
  * 会员卡相关接口
  *
- * @author YuJian(mgcnrx11@hotmail.com)
+ * @author YuJian(mgcnrx11@gmail.com)
  * @version 2017/7/8
  */
 public interface WxMpMemberCardService {
