+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/bean/membercard/MemberCardUserInfo.java
@@ -1,4 +1,4 @@
-package me.chanjar.weixin.mp.bean.mambercard;
+package me.chanjar.weixin.mp.bean.membercard;
 
 /**
  * Created by YuJian on 2017/7/11.
