+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/bean/membercard/WxMpMemberCardUserInfoResult.java
@@ -1,4 +1,4 @@
-package me.chanjar.weixin.mp.bean.mambercard;
+package me.chanjar.weixin.mp.bean.membercard;
 
 import java.io.Serializable;
