+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/util/json/WxMpMemberCardUserInfoResultGsonAdapter.java
@@ -2,9 +2,9 @@ package me.chanjar.weixin.mp.util.json;
 
 import com.google.gson.*;
 import me.chanjar.weixin.common.util.json.GsonHelper;
-import me.chanjar.weixin.mp.bean.mambercard.MemberCardUserInfo;
-import me.chanjar.weixin.mp.bean.mambercard.NameValues;
-import me.chanjar.weixin.mp.bean.mambercard.WxMpMemberCardUserInfoResult;
+import me.chanjar.weixin.mp.bean.membercard.MemberCardUserInfo;
+import me.chanjar.weixin.mp.bean.membercard.NameValues;
+import me.chanjar.weixin.mp.bean.membercard.WxMpMemberCardUserInfoResult;
 
 import java.lang.reflect.Type;
